+++ CHANGELOG.md
@@ -76,3 +76,4 @@ which block all requests data thread)
 ## 0.4.3
 
 * Update `BotLogger` behaviour
+* Fix problem with canceling of next scheduled job for `SchedulerPlugin`